Switched sample rate back to float. I suppose it's theoretically possible for there to be a non-integral number of samples per second and this is obvious when doing some calculations. (#60)

+++ src/main/kotlin/wrappers/sound/InputLine.kt
@@ -11,7 +11,7 @@ class InputLine(
     private val readSize: Int = 2048
 ) {
 
-    val sampleRate = dataLine.format.sampleRate.toInt()
+    val sampleRate = dataLine.format.sampleRate
     val bitDepth = dataLine.format.sampleSizeInBits
     val channels = dataLine.format.channels
     val byteOrder: ByteOrder = if (dataLine.format.isBigEndian) ByteOrder.BIG_ENDIAN else ByteOrder.LITTLE_ENDIAN
